Estimate networks from a multi-omics data set

coglasso() estimates multiple multi-omics networks with the algorithm collaborative graphical lasso, one for each combination of input values for the hyperparameters \(λ_w\), \(λ_b\) and \(c\).

Usage

coglasso(
  data,
  pX,
  lambda_w = NULL,
  lambda_b = NULL,
  c = NULL,
  nlambda_w = NULL,
  nlambda_b = NULL,
  nc = NULL,
  c_max = NULL,
  lambda_w_min_ratio = NULL,
  lambda_b_min_ratio = NULL,
  c_min_ratio = NULL,
  cov_output = FALSE,
  verbose = TRUE
)

Arguments

data

The input multi-omics data set. Rows should be samples, columns should be variables. Variables should be grouped by their assay (i.e. transcripts first, then metabolites). data is a required parameter.

pX

The number of variables of the first data set (i.e. the number of transcripts). pX is a required parameter.

lambda_w

A vector of values for the parameter \(λ_w\), the penalization parameter for the "within" interactions. Overrides nlambda_w.

lambda_b

A vector of values for the parameter \(λ_b\), the penalization parameter for the "between" interactions. Overrides nlambda_b.

c

A vector of values for the parameter \(c\), the weight given to collaboration. Overrides nc.

nlambda_w

The number of requested \(λ_w\) parameters to explore. A sequence of size nlambda_w of \(λ_w\) parameters will be generated. Defaults to 8. Ignored when lambda_w is set by the user.

nlambda_b

The number of requested \(λ_b\) parameters to explore. A sequence of size nlambda_b of \(λ_b\) parameters will be generated. Defaults to 8. Ignored when lambda_b is set by the user.

nc

The number of requested \(c\) parameters to explore. A sequence of size nc of \(c\) parameters will be generated. Defaults to 8. Ignored when c is set by the user.

c_max

The greatest generated \(c\). Defaults to 10. Ignored when c is set by the user.

lambda_w_min_ratio

The ratio of the smallest generated \(λ_w\) over the greatest generated \(λ_w\). Defaults to 0.1. Ignored when lambda_w is set by the user.

lambda_b_min_ratio

The ratio of the smallest generated \(λ_b\) over the greatest generated \(λ_b\). Defaults to 0.1. Ignored when lambda_b is set by the user.

c_min_ratio

The ratio of the smallest generated \(c\) over the greatest generated \(c\). Defaults to 0.1. Ignored when c is set by the user.

cov_output

Add the estimated variance-covariance matrix to the output.

verbose

Print information regarding current coglasso run on the console.

Value

coglasso() returns a list containing several elements:

loglik is a numerical vector containing the \(log\) likelihoods of all the estimated networks.

density is a numerical vector containing a measure of the density of all the estimated networks.

df is an integer vector containing the degrees of freedom of all the estimated networks.

convergence is a binary vector containing whether a network was successfully estimated for the given combination of hyperparameters or not.

path is a list containing the adjacency matrices of all the estimated networks.

icov is a list containing the inverse covariance matrices of all the estimated networks.

nexploded is the number of combinations of hyperparameters for which coglasso() failed to converge.

data is the input multi-omics data set.

hpars is the ordered table of all the combinations of hyperparameters given as input to coglasso(), with \(α(λ_w+λ_b)\) being the key to sort rows.

lambda_w is a numerical vector with all the \(λ_w\) values coglasso() used.

lambda_b is a numerical vector with all the \(λ_b\) values coglasso() used.

c is a numerical vector with all the \(c\) values coglasso() used.

pX is the number of variables of the first data set.

cov optional, returned when cov_output is TRUE, is a list containing the variance-covariance matrices of all the estimated networks.

Examples

# Typical usage: set the number of hyperparameters to explore
cg <- coglasso(multi_omics_sd_micro, pX = 4, nlambda_w = 3, nlambda_b = 3, nc = 3, verbose = FALSE)