Stability selection of the best coglasso network

stars_coglasso() selects the combination of hyperparameters given to coglasso() yielding the most stable, yet sparse network. Stability is computed upon network estimation from subsamples of the multi-omics data set, allowing repetition. Subsamples are collected for a fixed amount of times (rep_num), and with a fixed proportion of the total number of samples (stars_subsample_ratio).

Usage

stars_coglasso(
  coglasso_obj,
  stars_thresh = 0.1,
  stars_subsample_ratio = NULL,
  rep_num = 20,
  max_iter = 10,
  verbose = TRUE
)

Arguments

coglasso_obj

The object returned by coglasso().

stars_thresh

The threshold set for variability of the explored networks at each iteration of the algorithm. The \(\lambda_w\) or the \(\lambda_b\) associated to the most stable network before the threshold is overcome is selected.

stars_subsample_ratio

The proportion of samples in the multi-omics data set to be randomly subsampled to estimate the variability of the network under the given hyperparameters setting. Defaults to 80% when the number of samples is smaller than 144, otherwise it defaults to \(\frac{10}{n}\sqrt{n}\).

rep_num

The amount of subsamples of the multi-omics data set used to estimate the variability of the network under the given hyperparameters setting. Defaults to 20.

max_iter

The greatest number of times the algorithm is allowed to choose a new best \(\lambda_w\). Defaults to 10.

verbose

Print information regarding the progress of the selection procedure on the console.

Value

stars_coglasso() returns a list containing the results of the selection procedure, built upon the list returned by coglasso().

... are the same elements returned by coglasso().

merge_lw and merge_lb are lists with as many elements as the number of \(c\) parameters explored. Every element is in turn a list of as many matrices as the number of \(\lambda_w\) (or \(\lambda_b\)) values explored. Each matrix is the "merged" adjacency matrix, the average of all the adjacency matrices estimated for those specific \(c\) and \(\lambda_w\) (or \(\lambda_b\)) values across all the subsampling in the last path explored before convergence, the one when the final combination of \(\lambda_w\) and \(\lambda_b\) is selected for the given \(c\) value.

variability_lw and variability_lb are lists with as many elements as the number of \(c\) parameters explored. Every element is a numeric vector of as many items as the number of \(\lambda_w\) (or \(\lambda_b\)) values explored. Each item is the variability of the network estimated for those specific \(c\) and \(\lambda_w\) (or \(\lambda_b\)) values in the last path explored before convergence, the one when the final combination of \(\lambda_w\) and \(\lambda_b\) is selected for the given \(c\) value.

opt_adj is a list of the adjacency matrices finally selected for each \(c\) parameter explored.

opt_variability is a numerical vector containing the variabilities associated to the adjacency matrices in opt_adj.

opt_index_lw and opt_index_lb are integer vectors containing the index of the selected \(\lambda_w\)s (or \(\lambda_b\)s) for each \(c\) parameters explored.

opt_lambda_w and opt_lambda_b are vectors containing the selected \(\lambda_w\)s (or \(\lambda_b\)s) for each \(c\) parameters explored.

sel_index_c, sel_index_lw and sel_index_lb are the indexes of the final selected parameters \(c\), \(\lambda_w\) and \(\lambda_b\) leading to the most stable sparse network.

sel_c, sel_lambda_w and sel_lambda_b are the final selected parameters \(c\), \(\lambda_w\) and \(\lambda_b\) leading to the most stable sparse network.

sel_adj is the adjacency matrix of the final selected network.

sel_density is the density of the final selected network.

sel_icov is the inverse covariance matrix of the final selected network.

Details

StARS for collaborative graphical regression is an adaptation of the method published by Liu, H. et al. (2010): Stability Approach to Regularization Selection (StARS). StARS was developed for network estimation regulated by a single penalty parameter, while collaborative graphical lasso needs to explore three different hyperparameters. In particular, two of these are penalty parameters with a direct influence on network sparsity, hence on stability. For every \(c\) parameter, stars_coglasso() explores one of the two penalty parameters (\(\lambda_w\) or \(\lambda_b\)), keeping the other one fixed at its previous best estimate, using the normal, one-dimentional StARS approach, until finding the best couple. It then selects the \(c\) parameter for which the best (\(\lambda_w\), \(\lambda_b\)) couple yielded the most stable, yet sparse network.

Examples

cg <- coglasso(multi_omics_sd_micro, pX = 4, nlambda_w = 3, nlambda_b = 3, nc = 3, verbose = FALSE)
if (FALSE) {
# Takes around 20 seconds
sel_cg <- stars_coglasso(cg, verbose = FALSE)
}